Make layer number a short and Save 8 bytes.
By having a layer as a short (16 bits) we can pack it together with
the bitfields, saving 4 bytes per sub-struct, 8 bytes in total, also
bringing the struct down from 4 to 3 cachelines on my laptop.

Rationale: layers are mostly used to differentiate groups of objects
and they stacking, usually we have few layers and we use very large or
very small numbers to make a layer be at the top or at the bottom, but
usually we don't need so many layers.

Caution: code that use values like 999999 will break, so fix your
code! I'll provide another patch to fix all the CVS using these large
values.


SVN revision: 34420

+++ legacy/evas/src/lib/Evas.h
@@ -704,8 +704,8 @@ extern "C" {
 
    EAPI const char       *evas_object_type_get              (const Evas_Object *obj);
 
-   EAPI void              evas_object_layer_set             (Evas_Object *obj, int l);
-   EAPI int               evas_object_layer_get             (const Evas_Object *obj);
+   EAPI void              evas_object_layer_set             (Evas_Object *obj, short l);
+   EAPI short             evas_object_layer_get             (const Evas_Object *obj);
 
    EAPI void              evas_object_raise                 (Evas_Object *obj);
    EAPI void              evas_object_lower                 (Evas_Object *obj);
